[Journeys] Add click handler to navigate to visit urls
This is a small change on the MVC side of things, but requires some
moderate refactoring in order to cleanly wire the logic for building
the open url intent without depending on chrome_java or duplicating code
unnecessarily. Specifically, we wire the the intent creation logic
via a Function<> param to HistoryClustersCoordinator.

We also need to properly set the parent component for the intent that
launches the HistoryActivity for the Journeys UI; failing to do so means
we can mis-target the intent in multi-window.

Bug: 1303171
Change-Id: Ic7da8548009f7988bb72b4640f7f137875ba1ada
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3642523
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002899}