Move the Host Preprocessing into the Page Topics Executor
Having the preprocessing logic in the PCAService is cumbersome:
* It'll make integration testing harder since the testing input would
also have to be preprocessed, which isn't a good test.
See https://crrev.com/c/3635390
* Callers have to keep track of the original input, which also sucks.

The logic is now moved to the executor itself, just before passing it to
the BERT executor, and the original input is passed back to the caller.

Bug: 1288877
Change-Id: I4633a23af28246a39a7d3e1bd1511d41655384c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3635525
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002919}