[CrOS SLP] Surface notification on newly active user sessions.
Surface notification during the start of each newly active user session
if the active SIM is locked and the policy is set to true.

I.e if user dismisses the notification without taking action, locks the
device and unlocks the device, or logs out and logs back in, the
notification will surface again until the SIM is unlocked.

Bug: b/228093904
Change-Id: I79faa5cef12ff86f6a261a00d7dcdfc6cbffaa79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3640955
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002922}