Apply mobile-style Blink prefs to the New Tab Page
Since 2018, Chrome on ChromeOS would update its Blink preferences to
emulate a mobile browser when in tablet mode. [1]

Tabs that were initialized to the New Tab Page wouldn't have their Blink
preferences updated to emulate a mobile browser when in tablet mode.
This was done because it was understood that internal system pages
already work well in a tablet touchscreen context and do not need
the browser to intervene using Blink features like fit-to-width and
text autosizing. [2]

A likely unintended consequence of this exclusion is that any website
visited within that same new tab also doesn't get the mobile-style Blink
preferences. This is because the Blink preferences are not updated when
the URL changes -- they're only updated when the tab is created and when
tablet mode is toggled. The result is that mobile-style Blink
preferences often aren't applied to sites visited in new tabs even when
in tablet mode.

A straightforward way to mitigate this problem in the common case is to
not exclude the New Tab Page from having mobile-style Blink preferences.
As far as I can tell, this doesn't negatively impact the touch usability
of the New Tab Page.

[1] d70bf37b08c43fa8307aae7fd90086bad16e5a53
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=835939#c8

Bug: 1317927
Change-Id: I549eaad33d3b958c10fed0a41727ab908938b9cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3646553
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Alex Yang <aycyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002944}