MPArch: Convert DidFinishNavigation to PrimaryPageChanged in PasswordGenerationPopupControllerImpl
As a clean-up of fixing long tail of features for MPArch, this CL
renames DidFinishNavigation to PrimaryPageChanged in
PasswordGenerationPopupControllerImpl, because PrimaryPageChanged
is preferred when DidFinishNavigation only handles changes to the
document.

And also, the MPArch test guideline allows converting to
PrimaryPageChanged without a test when the code is simple.

Bug: 1269645
Change-Id: I6a1749f76eb5be6261c83855076d38b65ad233e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3641195
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002975}