MPArch: Convert DidFinishNavigation to PrimaryPageChanged in SideSearchSideContentsHelper
As a clean-up of fixing long tail of features for MPArch, this CL
renames DidFinishNavigation to PrimaryPageChanged in
SideSearchSideContentsHelper, because PrimaryPageChanged
is preferred when DidFinishNavigation only handles changes to the
document.

And also, the MPArch test guideline allows converting to
PrimaryPageChanged without a test when the code is simple.

Bug: 1269645
Change-Id: Ic585d4d9acacb816b3d36671ab955a3b445a7006
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3640590
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002979}