[views] Eliminate OnWidgetClosing from c/b/ui/views/autofill/payments
Favor OnWidgetDestroying over OnWidgetClosing. OnWidgetClosing is
only called where a Widget is closed directly, it fails to catch
the case when the system destroys the NativeWidget directly. This
can result in UAFs in situations where the OS directly destroys
the widget.

OnWidgetDestroying will be called every time a widget is destroyed
and should be used instead in most cases.

This CL adds the GetPaymentsBubbleClosedReasonFromWidget() utility
that is used to extract the correct PaymentsBubbleClosedReason.
DCHECKs on close conditions have been moved to OnWidgetDestroying()
if necessary.

Bug: 1240365
Change-Id: I889d21464eb3fefba459ec7b58e3e31d3b1e8103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3642558
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002991}