emoji-picker: add symbols data generator script
A script is implemented to create symbol data. The data includes
symbol groups and a set of symbols for each one of them. The symbols
are generated based on a set of position ranges from standard unicode
blocks that are conceptually relevant to the group. More details can
be found at go/emoji-picker-symbol-data-doc.

Bug: b:232160008
Change-Id: I113dc514e3f3f10a1d5b4d4d293a3f328f7442a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3640589
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Mehrab N <mehrab@chromium.org>
Reviewed-by: Michael Cui <mlcui@google.com>
Reviewed-by: John Palmer <jopalmer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002998}