crostini_upgrader: Fix crash if launch_callback_ is null
::Launch assumes that launch_callback_ is set and runs it, and two of
the call sites in the same file check this before calling, but it is
also a mojo method, so it can be called from the renderer process
without this check.

This could cause a (deterministic) crash if e.g. the WebUI called
Launch twice, or if it called Launch and closed the page, and the
::OnPageClosed event was handled first.

Fix this by moving the check inside ::Launch

Bug: b:232481454
Change-Id: Ic603fce1235d0da1a8fdf834d946ff127a7d3788
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3647272
Commit-Queue: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003029}