Use only a single CrostiniRestarter per ContainerId
Currently we create a CrostiniRestarter for every restart request, but
only run the first one unless it is aborted (incl. via options like
start_vm_only). This CL changes the model to use only a single restarter
object per ContainerId, which is easier to reason about and less
error-prone.


Details:

- Primarily a refactoring change, but there are several edge case bug
fixes described below.

- CrostiniRestarter now stores a vector of "RestartRequest" objects,
which are the restart id, callback, options, and observer for a given
request.

- Aborting restarters has been split into cancelling requests and
aborting restarters.
-- CrostiniRestarter::CancelRequest() is for external clients, right now
only the installer. If other requests are present, this doesn't stop the
restarter and allows them to continue (options like start_vm_only behave
similarly). As a result, observers can now assume that states always
progress in order. The installer already assumes this so this probably
fixes a crash when sharing a folder after a failed install and then
retrying the install (see b/230670356).
-- CrostiniRestarter::Abort() is used internally by CrostiniManager for
the uninstall flow. Abort callbacks are now called in FinishRestart()
instead of ReturnEarlyIf..(), which fixes an issue where uninstall could
get stuck if (e.g.) a restarter timed out as the abort callback wouldn't
get called. The new Abort() exists as a way to have a callback fire once
the restarter is done, which doesn't make sense for CancelRequest().

- CrostiniRestarter::FinishRestart now always deletes the object. This
probably fixes issues if the container is running debian stretch and the
restart timed out (the restarter is currently not immediately deleted,
so there will be problems if the timed out stage actually completes or
the restarter is cancelled).

- Fixes handling of CrostiniRestarter::observer_list_. Currently when
adding a new observer it is added to all restarters, but when removed it
is only removed from the directly associated restarter, which probably
leads to crashes or worse in some cases (e.g. open installer -> install
failure -> set up termina via vmc -> begin terminal launch -> retry
install -> cancel install at appropriate time).

Bug: b/227693911
Change-Id: I5c5cb8576c0e003c12ee5b13010fbb32455cd24d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3583822
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003035}