Move logging to consistent place
Before this change the logging callback was called in
two places: In the `HandleX` method and in the
`ReloadUserData` method.

Now the call only happens in the `HandleX` method for
consistency. It does not change how each event is logged,
nor does it change the timing.

Bug: b:230844184
Change-Id: I6828416d60ca18307c98f816e88f9ec3853f6fe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3644733
Reviewed-by: Luca Hunkeler <hluca@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/main@{#1003110}