[DownloadBubble] Limit number of downloads shown on the bubble
Limit the downloads shown on the bubble to a maximum of 100. This
will prevent OOM crashes happening.

The value of 100 is chosen after looking at the
Download.Bubble.FullViewSize and Download.Bubble.PartialViewSize
metrics. For the canary/dev channels, 95 and 99 percentiles for the
partial view size are ~6 and 58, and for the main view are 50 and
100 (max). Even in the unlikely case, the user is able to go to
chrome://downloads for the full list of all downloads.

Bug: 1282247
Bug: 1315554
Bug: 1320415
Change-Id: I76680cc7c9cb0d7bd542c0f2a616123ffefaaab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3646559
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Auto-Submit: Rohit Bhatia <bhatiarohit@google.com>
Cr-Commit-Position: refs/heads/main@{#1003158}