Reduce the number of shards for ChromeOS MSAN bot
A portion of the recent slowdown on MSAN has been addressed, so we can
now reduce the number of shards used on the MSAN bot. The fully
instrumented ANGLE + SwiftShader Vulkan libraries offer a more thorough
coverage, but are heavier than legacy SwiftShader was, so instead of
going back to the original 48 shards, we're going back half way to 60
shards.

Bug: chromium:1313907
Change-Id: Ic56478708ffb61117866982e93b10a4fb4699990
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645972
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@google.com>
Cr-Commit-Position: refs/heads/main@{#1003167}