[bfcache] Include cache-control:no-store reasons in tree
This CL adds cache-control:no-store reasons in the tree structure, as
well as the flattened list. In order to achieve it, this CL starts to
mark CCNS related reasons all the time in NotRestoredReasons, and
introduce CanStore() and CanRestore().

CanStore() considers CCNS as blocking only when other reasons are
recorded as well, where as CanRestore() considers CCNS always blocking.

Bug: 1280150
Change-Id: I33af681aef64056d9d92d12fdf0942dfe57411d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3610704
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003168}