Use Python 3 from ./PRESUBMIT.py
The root directory PRESUBMIT.py is run using Python 3, but it invokes
other scripts and it was running two of them using python_executable
instead of python3_executable. This change fixes that.

checkperms.py had to be updated because the filter command returns an
iterable object instead of a list. Converting it to a list ensures that
an empty list will be False, as opposed to the object which is always
True. I fixed this on one code path a year ago (crrev.com/c/2806723)
but now the fix covers all paths.

This was tested (on Linux) with the following command:

    git cl presubmit --files=PRESUBMIT.py

Bug: 1313804
Change-Id: I1c8a0d1e5e888b9e7ee6cfcbeb218b0fb9aaad6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3643458
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003172}