Move LacrosFirstRun helpers to a KeyedService
The goal with this is to simplify the management of the lifecycle of
instances that will be needed for further features, and to use a more
standard way to trigger First Run checks on startup.


Another change introduced in this CL is the suppression of the FRE for
Guest sessions and non-GAIA ones.

Bug: 1300109,1324569
Change-Id: Ie8f7391d1e4a3a7b6873ae4664b30218fc062435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3647050
Auto-Submit: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003182}