[Feature Discovery] Support to record data under the specified mode
This CL allows users to pass two histogram strings, one for clamshell
and the other for tablet, when constructing a `TrackableFeatureInfo`
instance. When such a constructor is used, it means that the metric data
collected should be separated by the mode, i.e. clamshell or tablet,
under which the feature observation starts.

`FeatureDiscoveryDurationReporter` saves a boolean value that indicates
the current mode (tablet or clamshell) when the observation is
activated. Then when the observation completes, the reporter fetches
this boolean value from the pref service then uses this boolean to
decide which histogram should be used for reporting.

Bug: 1311344
Change-Id: I938ef0e5d8de2afb1e63a4ccfcea850ed2e64d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3636360
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003191}