Change quota and usage to be calculated based on storage key rather than host.
This does change the behavior even when third party storage partitioning is
not turned on, since quota and usage are now calculated based on origin
rather than just the host name. This is what most people seem to assume
it was already anyway, so if anything this should reduce the number of
surprising situations where for example the usage estimate API shows a
site using a bunch of data while none of the storage APIs actually have
data.

GetHostUsageWithBreakdown still exists as BrowsingDataQuotaHelperImpl
still uses it to gather information for CookiesTreeModel. That code
will be updated to also get information at a StorageKey level instead
in a follow-up.

Persistent quota is still shared per host as well, the deprecation of
persistent quota as separate from temporary quota will fix that
issue.

Bug: 1321895
Change-Id: I2400d2f2e2e13b9324db82249687e4300dda33a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645668
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003210}