DanglingPtr: Add an option to list dangling ptr.
To track progress, I would need a way to list all the tasks. That is to
say, every occurrences of dangling pointers. I would need to produce tables
similar to this one:
https://docs.google.com/spreadsheets/d/1BAfz1uHhMkFe2sK2vyTHsJ733VQUhY9v2M4HxOwaM7E/edit?resourcekey=0-H9SwkJTMbBxQPKUEuIrwow#gid=0

Behind the feature flag "LogDanglingPtr", this patch turns crash into logging the signature of the occurrence. From the StackTraces, the
relevant callers of free() and raw_ptr.reset() are extracted. Those
are the "signature" of the event. This is easier to process,
compared to the full pair of StackTraces.

This patch add the option "LogDanglingPtr", which makes Chrome to stop
crashing when it encounter a dangling raw_ptr. Instead it prints the
signature of the dangling ptr::
- where it was memory was `free()`.
- where the dangling raw_ptr was released.

A documentation in `docs/dangling_ptr.md` is created.

Bug: 1291138
Change-Id: I05f08dc1c6ab62f82243cadbb8e862260ec54c98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3575947
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003214}