[Topics] topics-internals: display a warning on invalid classifier input
What: In the "Classifier" tab, display a warning if the input string
has a '/' in it.

Why: the user may attempt to type full URLs, but the classifier is only
expecting hostnames. The result topics can appear to be wrong due to
the misuse and misunderstanding.

See https://crbug.com/1325402#c1 for screenshots.

Bug: 1325402
Change-Id: If3bed831d14c54531b872bef04d05f000ddbec03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3647724
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003269}