[floats] Add logic to ensure we don't intersect floats.
Previously with certain values of margins - it was possible to create
a scenario to intersect a new-fc with a float.

As demonstrated by: floats-wrap-bfc-with-margin-003.tentative.html

This patch introduces logic to prevent this. Specifically after we have
calculated the BFC-offset for the new formatting context, we check
if it is outside the opportunity on each side, and if that side of the
opportunity has any floats.

Bug: 1323004, 1322774
Change-Id: I82e07e6568da92a3e73842f54347c4f07ec73f51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3636486
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003290}