lacros: Refactor fileSystemProvider functions to use mojom.
This CL is a refactor with no intended behavior change.

In preparation for migration to Lacros, this CL creates a new crosapi
FileSystemProvider. This interface currently mirrors the public
interface of the extension API fileSystemProvider. The ash-specific
parts of the implementation of fileSystemProvider is moved into
FileSystemProviderServiceAsh. This CL adds a mojom translation layer.

Change-Id: I57e62366d10a7d84552c1c650411620ef44a3cb1
Bug: 1314397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3636237
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003292}