Window state restore history: Remove kPreMinimizedShowStateKey property.
- Now we have window state restore history stack, we can get the state
type to restore back to from the restore history stack, so we no longer
need to maintain this window property. Same thing for
kPreFullscreenShowStateKey, which will be removed in a following CL.

- kRestoreShowStateKey is introduced as a way to communicate the desired
window state to restore back to between ash and other components. It's
to replace kPreMinimizedShowStateKey and kPreFullscreenShowStateKey
window properties. And because of this new window properly,
kIsRestoringKey can also be removed.

No functional change in this CL.

TODO:
- I'll remove kPreFullscreenShowStateKey in a following CL.

Bug: 990494, 939031
Change-Id: I1f7be0da9899d498a1544c1c49a65b0d11434d08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3640222
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003325}