[Chromecast] Timing-based methods for Observable.
This adds two methods:

* Observable.alarm() activates once after a given delay.
* Observable.delay() delays the activations of the calling
  Observable by the given delay.

These methods are agnostic about the mechanism to delay work for
purposes of testing and of avoiding dependencies on the Android
sdk. Therefore, these methods require injecting a delegate that
is responsible for running delayed tasks on the right thread.

To make this easy in Android code, this patch also provides a
HandlerScheduler that does the right thing using Android
Handlers.

With these tools, the result of:

  src.delay(HandlerScheduler.onCurrentThread(), 200)

... will be an Observable that yields the activations of |src|,
but only 200 milliseconds after they are activated in |src|.

This is useful for "debouncing" Observables, among other things.

This is added specifically for a situation in chromecast
internal code. See the linked bug for details.

Bug: b/219101568
Test: cast_base_junit_tests, cast_shell_junit_tests
Change-Id: If9f52686affd3eeea63673e0efd2d41586e95cb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3636364
Commit-Queue: Simeon Anfinrud <sanfin@chromium.org>
Commit-Queue: Yuchen Liu <yucliu@chromium.org>
Auto-Submit: Simeon Anfinrud <sanfin@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003328}