[Shill] Update shill manager
This CL updates FakeShillManagerClient, When a devices technology
state is updated to initializing, the device is still in the list of
enabled devices, this means if a check is made for the current
device state, the device  will be reported as being enabled. This
causes issues when testing. This CL fixes that issue, it removes a
device from the list of enabled devices and notifies of this change,
it also adds a new function to update a technology to disabling state.
This new function will be used in new NetworkListViewController test.

bug: b/228497990
Test: Ran CQ
Change-Id: Ib2d273d4aa0584a39e75bac5f328d1bfc05088fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3643216
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Commit-Queue: Theo Johnson-kanu <tjohnsonkanu@google.com>
Cr-Commit-Position: refs/heads/main@{#1003353}