Quota: Use BindPostTask in QuotaManagerProxy
This change updates QuotaManagerProxy to use
base::BindPostTask to ensure callbacks are run or
destroyed on the specified task runner. The main difference
between the existing logic is that if the callback
is never run, it may not be destroyed on the correct task
runner. base::BindPostTask makes sure this doesn't happen.

Bug: 1324864
Change-Id: Ifd123031309e29a22761b8aa88ef2a216ad45c70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3646864
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003354}