[Jank] Add TraceEvents to many WebContentsObservers
FrameHost is the second most significant contributor according to
go/scroll-jank-effort. The top contributors to jank are from notifying
WebContentsObservers during navigations events. We want to see which
specific observers are slow.

Perf sherriffs: if this triggers small regressions, we still want to
ship a few builds to canary/dev to collect data. If it's egregious
then ok to revert


Bug: 1325438
Change-Id: I6651badead97f93a8624816473448833b807a89e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645631
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1003416}