[OOBE] Handle net::ERR_ABORTED in TOS webview
When a webview is reloading, net::ERR_ABORTED error is thrown to stop
the previous request and start a new one.

Currently we show the error screen without checking the kind of error
which results in error being shown if we retry or the system attempts to
load/refresh the UI automatically soon after initiating it earlier.

This is a temporary fix. A more comprehensive fix would require that we
keep track of current state of request and only initiate a new one when
previous one has completed or failed with an error. A similar approach
is used by oobe_eula.js

Bug: b:232177321
Bug: b:232592745
Test: deploy to DUT, reset to OOBE flow, login and validate TOS load
Change-Id: I2a4031f19a4297ab30a93d4eaa64fb63db5f7033
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645553
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003458}